refactor(uilanding): extract page and topology API handlers with tests
- Extract 9 page/utility handlers into handlers/pages.py
- Extract 6 topology/device API handlers into handlers/topology_api.py
  (including TopologyAPIHandler ~800 lines, device utility functions)
- Add test_pages.py (12 tests) and test_topology_api.py (36 tests)
- 271 total tests passing
