+++ nested-labvm/atd-docker/uilanding/src/uilanding.py
@@ -73,6 +73,29 @@ from handlers.impairments import (
     LatencyDisableAPIHandler,
     LatencyDisableAllAPIHandler,
 )
+from handlers.pages import (
+    topoRequestHandler,
+    ToolsHandler,
+    ViewConfigHandler,
+    BaseUrlHandler,
+    UptimeWithRuntimeHandler,
+    TerminalPageHandler,
+    ConsolePageHandler,
+    ClientLogHandler,
+    ConnectivityStatusHandler,
+)
+from handlers.topology_api import (
+    TopologyAPIHandler,
+    DevicesAPIHandler,
+    DeviceTypesAPIHandler,
+    InterfaceStatsAPIHandler,
+    DeviceStatusAPIHandler,
+    RunningConfigAPIHandler,
+    get_all_devices,
+    get_device_ip_from_sources,
+    invalidate_devices_cache,
+    initialize as initialize_topology_api,
+)
 
 # Disable any TLS Warnings when getting instance Uptime
 urllib3.disable_warnings()
@@ -164,6 +187,10 @@ elif 'title' in host_yaml:
 else:
     TITLE = 'Arista Training Lab'
 
+# Propagate config to topology_api module (called after all globals are resolved)
+initialize_topology_api(TOPO, EOS_TYPE, TITLE, ATD_ACCESS_PATH)
+
+
 def get_metadata_extract(attribute):
     try:
         metadata_url = "http://169.254.169.254/computeMetadata/v1/project/attributes/{}".format(attribute)
@@ -181,98 +208,8 @@ HonorLockClientID = get_metadata_extract('honorlockClientID')
 HonorLockSecret = get_metadata_extract('honorlockClientSecret')
 
 
-class topoRequestHandler(BaseHandler):
-    def get(self):
-        with open(ATD_ACCESS_PATH, 'r') as f:
-            host_yaml = YAML().load(f)
-        lab_type = host_yaml.get('customer_details', {}).get('lab_type', 'Lab')
+# topoRequestHandler — moved to handlers/pages.py
 
-        # For Exam labs: ALWAYS require Honorlock authentication flow
-        # This prevents bypass via direct URL access or cached sessions
-        if lab_type == "Exam":
-            # Check if user accessed via proper Honorlock flow (has 'honorlock' parameter)
-            if 'honorlock' in self.request.arguments:
-                # Valid Honorlock flow - proceed with authentication
-                if not self.current_user:
-                    # Redirect to login with auth credentials
-                    if 'auth' in self.request.arguments:
-                        self.redirect('/login?auth={0}'.format(self.get_argument('auth')))
-                    else:
-                        self.redirect('/login')
-                    return
-                # else: User authenticated via Honorlock - allow access (continue to line 232+)
-            else:
-                # No 'honorlock' parameter - force Honorlock authentication
-                # This blocks:
-                #   1. Direct URL access: https://lab.com/
-                #   2. Cached session access without Honorlock
-                #   3. Manual auth parameter manipulation
-                self.redirect('/exam-authentication')
-                return
-
-        # Handle non-authenticated users for regular (non-Exam) labs
-        # Note: Exam labs are already handled above (lines 210-229)
-        if not self.current_user:
-            # Regular lab authentication
-            if 'auth' in self.request.arguments:
-                self.redirect('/login?auth={0}'.format(self.get_argument('auth')))
-            else:
-                self.redirect('/login')
-            return
-        else:
-            safe_log('info', 'Topology page accessed', event='page_view', page='topology', lab_type=str(lab_type))
-            _topo_cvp = False
-            if 'disabled_links' in host_yaml:
-                disable_links = host_yaml['disabled_links']
-            else:
-                disable_links = []
-            menu={}
-            if NOMENUOPTIONFILE:
-                disable_links.append('lab_menu')
-            else:
-                for lab in MENU_ITEMS['lab_list']:
-                    menu[lab] = MENU_ITEMS['lab_list'][lab]['description']
-
-            # Disable lab_menu for Exam type labs
-            if lab_type == "Exam" and 'lab_menu' not in disable_links:
-                disable_links.append('lab_menu')
-            if 'labguides' in host_yaml:
-                if host_yaml['labguides'] == 'self':
-                    labguides = '/labguides/index.html'
-                else:
-                    labguides = host_yaml['labguides']
-            else:
-                labguides = '/labguides/index.html'
-            if 'cvp' in host_yaml:
-                if host_yaml['cvp'] != "none":
-                    _topo_cvp = True       
-            gui_urls,servers =[],[]
-            if host_yaml['eos_type'] == 'container-labs':
-                try:
-                    servers =  MOD_YAML['topology']['servers']
-                    if servers is None:
-                        servers = [] 
-                    external_ip_url = "http://metadata.google.internal/computeMetadata/v1/instance/network-interfaces/0/access-configs/0/external-ip"
-                    headers = {"Metadata-Flavor": "Google"}
-                    response = requests.get(external_ip_url, headers=headers, timeout=5)
-                    for server in servers:
-                        gui_urls.append(f'http://{response.text}:{servers[server]["port"]}')
-                except Exception as e:
-                    safe_log('error', f'Error in topoRequestHandler: {e}', event='error', handler='topoRequestHandler')
-            self.render(
-                BASE_PATH + 'index.html',
-                NODES = MOD_YAML['topology']['nodes'],
-                SERVERS = servers,
-                GUI_URLS= gui_urls,
-                ARISTA_PWD=host_yaml['login_info']['jump_host']['pw'],
-                topo_title = TITLE,
-                disable_links = disable_links,
-                labguides = labguides,
-                topo_cvp = _topo_cvp,
-                menu_options = menu,
-                lab_type = lab_type
-            )
-    
 # ===============================
 # Internal CVP gRPC Health Check
 # ===============================
@@ -316,11 +253,10 @@ def check_cvp_grpc_internal():
     Uses same framing and auth as the frontend check.
     Results logged to Cloud Logging for comparison with frontend reports.
     """
-    global _last_internal_grpc_status, _last_internal_grpc_time
     token = _get_cvp_token()
     if not token:
-        _last_internal_grpc_status = 'skipped'
-        _last_internal_grpc_time = time.time()
+        _grpc_state['status'] = 'skipped'
+        _grpc_state['last_check'] = time.time()
         safe_log('warning', 'Internal gRPC check skipped - no token',
             event='connectivity', action='grpc_check', source='internal',
             status='skipped', reason='no_token')
@@ -352,13 +288,13 @@ def check_cvp_grpc_internal():
         if status_code == 200 or grpc_status is not None:
             # Check grpc-status if present
             if grpc_status is not None and int(grpc_status) == 14:
-                _last_internal_grpc_status = 'unavailable'
+                _grpc_state['status'] = 'unavailable'
                 safe_log('warning', 'Internal gRPC check: CVP unavailable',
                     event='connectivity', action='grpc_check', source='internal',
                     status='unavailable', http_status=str(status_code),
                     grpc_status=str(grpc_status))
             else:
-                _last_internal_grpc_status = 'ok'
+                _grpc_state['status'] = 'ok'
                 safe_log('info', 'Internal gRPC check passed',
                     event='connectivity', action='grpc_check', source='internal',
                     status='ok', http_status=str(status_code),
@@ -367,30 +303,30 @@ def check_cvp_grpc_internal():
             # Token may be stale, clear it
             global _cvp_grpc_token
             _cvp_grpc_token = None
-            _last_internal_grpc_status = 'auth_rejected'
+            _grpc_state['status'] = 'auth_rejected'
             safe_log('warning', 'Internal gRPC check: auth rejected',
                 event='connectivity', action='grpc_check', source='internal',
                 status='auth_rejected', http_status=str(status_code))
         elif status_code in (502, 503, 504):
-            _last_internal_grpc_status = 'unreachable'
+            _grpc_state['status'] = 'unreachable'
             safe_log('warning', 'Internal gRPC check: CVP unreachable',
                 event='connectivity', action='grpc_check', source='internal',
                 status='unreachable', http_status=str(status_code))
         else:
-            _last_internal_grpc_status = 'unexpected'
+            _grpc_state['status'] = 'unexpected'
             safe_log('warning', 'Internal gRPC check: unexpected response',
                 event='connectivity', action='grpc_check', source='internal',
                 status='unexpected', http_status=str(status_code))
-        _last_internal_grpc_time = time.time()
+        _grpc_state['last_check'] = time.time()
     except requests.exceptions.Timeout:
-        _last_internal_grpc_status = 'timeout'
-        _last_internal_grpc_time = time.time()
+        _grpc_state['status'] = 'timeout'
+        _grpc_state['last_check'] = time.time()
         safe_log('warning', 'Internal gRPC check: timeout',
             event='connectivity', action='grpc_check', source='internal',
             status='timeout')
     except Exception as e:
-        _last_internal_grpc_status = 'error'
-        _last_internal_grpc_time = time.time()
+        _grpc_state['status'] = 'error'
+        _grpc_state['last_check'] = time.time()
         safe_log('error', 'Internal gRPC check failed',
             event='connectivity', action='grpc_check', source='internal',
             status='error', error=str(e))
@@ -404,8 +340,8 @@ def check_cvp_grpc_internal():
 recent_sessions = {}
 active_sessions = set()  # Set of session IDs currently connected
 active_session_data = {}  # Snapshot of session info keyed by session_id
-_last_internal_grpc_status = None
-_last_internal_grpc_time = None
+# Mutable dict so ConnectivityStatusHandler can hold a reference and see live updates
+_grpc_state = {'status': None, 'last_check': None}
 RECONNECT_WINDOW_SECONDS = 300  # 5 minutes
 
 def prune_recent_sessions():
@@ -577,8 +513,8 @@ class topoDataHandler(tornado.websocket.WebSocketHandler):
             # Send timestamped ping for latency measurement
             # Include internal gRPC status when available for synchronized checks
             ping_data = {'ts': int(time.time() * 1000)}
-            if _last_internal_grpc_status is not None:
-                ping_data['internal_grpc'] = _last_internal_grpc_status
+            if _grpc_state['status'] is not None:
+                ping_data['internal_grpc'] = _grpc_state['status']
             if hasattr(self, 'session') and self.session['last_rtt'] is not None:
                 ping_data['server_rtt'] = self.session['last_rtt']
             self.write_message(json.dumps({
@@ -814,293 +750,8 @@ class topoDataHandler(tornado.websocket.WebSocketHandler):
                 self.schedule_summary()
 
 
-# Cache for topo_build.yml data (loaded once on first use)
-_TOPO_BUILD_CACHE = None
-# Cache for merged device list from both sources
-_ALL_DEVICES_CACHE = None
-# Track user_nodes.yaml modification time for cache invalidation
-_USER_NODES_MTIME = 0
-# Track user_hosts.yaml modification time for cache invalidation
-_USER_HOSTS_MTIME = 0
-# Track user_firewalls.yaml modification time for cache invalidation
-_USER_FIREWALLS_MTIME = 0
-# Track user_velo.yaml modification time for cache invalidation
-_USER_VELO_MTIME = 0
-
-
-def invalidate_devices_cache():
-    """
-    Invalidate the devices cache.
-    Called when user nodes/hosts/firewalls/velo devices are added/removed to ensure fresh data.
-    """
-    global _ALL_DEVICES_CACHE, _USER_NODES_MTIME, _USER_HOSTS_MTIME, _USER_FIREWALLS_MTIME, _USER_VELO_MTIME
-    _ALL_DEVICES_CACHE = None
-    _USER_NODES_MTIME = 0
-    _USER_HOSTS_MTIME = 0
-    _USER_FIREWALLS_MTIME = 0
-    _USER_VELO_MTIME = 0
-    pS("Devices cache invalidated")
-
-
-def _get_topo_build_data():
-    """
-    Load and cache topo_build.yml data.
-    Returns cached data on subsequent calls.
-    """
-    global _TOPO_BUILD_CACHE
-
-    if _TOPO_BUILD_CACHE is not None:
-        return _TOPO_BUILD_CACHE
-
-    topo_path = f"/opt/atd/topologies/{TOPO}/topo_build.yml"
-    try:
-        with open(topo_path, 'r') as f:
-            _TOPO_BUILD_CACHE = YAML().load(f)
-        pS(f"Cached topo_build.yml from {topo_path}")
-    except Exception as e:
-        safe_log('error', f'Error in _get_topo_build_data: {e}', event='error', handler='_get_topo_build_data')
-        _TOPO_BUILD_CACHE = {}  # Empty dict to avoid repeated failures
-
-    return _TOPO_BUILD_CACHE
-
-
-def get_device_ip_from_sources(device_name):
-    """
-    Look up device IP using cached get_all_devices() with case-insensitive matching.
-
-    Args:
-        device_name: Name of the device to look up
-
-    Returns:
-        str: IP address if found, None otherwise
-    """
-    if not device_name:
-        return None
-
-    all_devices = get_all_devices()
-    device_name_lower = device_name.lower()
-
-    # Case-insensitive lookup in cached devices
-    for name, info in all_devices.items():
-        if name.lower() == device_name_lower:
-            ip = info.get('ip', '')
-            return ip if ip else None
-
-    return None
-
-
-def normalize_device_name(name):
-    """
-    Normalize device name to consistent capitalization.
-    E.g., 'leaf5' -> 'Leaf5', 'memleaf1' -> 'Memleaf1', 'spine1-dc1' -> 'Spine1-DC1'
-
-    Handles patterns like:
-    - Simple: leaf1 -> Leaf1, spine2 -> Spine2
-    - With suffix: spine1-dc1 -> Spine1-DC1, leaf2-dc2 -> Leaf2-DC2
-    - Compound: memleaf1 -> Memleaf1, borderleaf1 -> Borderleaf1
-    - Abbreviations: PE1 -> PE1, P3 -> P3 (preserve uppercase abbreviations)
-    """
-    import re
-
-    if not name:
-        return name
-
-    # Split on hyphens to handle suffixes like -DC1, -DC2
-    parts = name.split('-')
-    result_parts = []
-
-    for part in parts:
-        # Check if this part is a datacenter suffix (DC1, DC2, etc.)
-        if re.match(r'^[dD][cC]\d+$', part):
-            # Uppercase the DC suffix
-            result_parts.append(part.upper())
-        # Check if this part is an uppercase abbreviation followed by numbers (PE1, P3, CE1, etc.)
-        elif re.match(r'^[A-Z]+\d*$', part):
-            # Preserve uppercase abbreviations like PE1, P3, CE1
-            result_parts.append(part)
-        else:
-            # Capitalize first letter, keep rest of case
-            # This turns 'leaf5' -> 'Leaf5', 'memleaf1' -> 'Memleaf1'
-            result_parts.append(part.capitalize())
-
-    return '-'.join(result_parts)
-
-
-def get_all_devices():
-    """
-    Get all devices from topo_build.yml, user_nodes.yaml, user_hosts.yaml, and user_firewalls.yaml.
-    Returns a dict of {device_name: {'ip': ip_address, 'user_added': bool, 'device_category': str}}.
-    Uses caching to avoid repeated lookups.
-    Device names are normalized to consistent capitalization.
-    Cache is auto-invalidated when any user file changes.
-    """
-    global _ALL_DEVICES_CACHE, _USER_NODES_MTIME, _USER_HOSTS_MTIME, _USER_FIREWALLS_MTIME, _USER_VELO_MTIME
-
-    # Check if any user file has been modified since last cache
-    user_nodes_path = '/etc/atd/user_nodes.yaml'
-    user_hosts_path = '/etc/atd/user_hosts.yaml'
-    user_firewalls_path = '/etc/atd/user_firewalls.yaml'
-    user_velo_path = '/etc/atd/user_velo.yaml'
-
-    try:
-        if os.path.exists(user_nodes_path):
-            current_mtime = os.path.getmtime(user_nodes_path)
-            if current_mtime > _USER_NODES_MTIME:
-                _ALL_DEVICES_CACHE = None
-                _USER_NODES_MTIME = current_mtime
-    except OSError:
-        pass
-
-    try:
-        if os.path.exists(user_hosts_path):
-            current_mtime = os.path.getmtime(user_hosts_path)
-            if current_mtime > _USER_HOSTS_MTIME:
-                _ALL_DEVICES_CACHE = None
-                _USER_HOSTS_MTIME = current_mtime
-    except OSError:
-        pass
-
-    try:
-        if os.path.exists(user_firewalls_path):
-            current_mtime = os.path.getmtime(user_firewalls_path)
-            if current_mtime > _USER_FIREWALLS_MTIME:
-                _ALL_DEVICES_CACHE = None
-                _USER_FIREWALLS_MTIME = current_mtime
-    except OSError:
-        pass
-
-    try:
-        if os.path.exists(user_velo_path):
-            current_mtime = os.path.getmtime(user_velo_path)
-            if current_mtime > _USER_VELO_MTIME:
-                _ALL_DEVICES_CACHE = None
-                _USER_VELO_MTIME = current_mtime
-    except OSError:
-        pass
-
-    if _ALL_DEVICES_CACHE is not None:
-        return _ALL_DEVICES_CACHE
-
-    devices = {}
-
-    # Get devices from topo_build.yml (the authoritative topology source)
-    topo_data = _get_topo_build_data()
-    if topo_data and 'nodes' in topo_data:
-        for node_entry in topo_data['nodes']:
-            if isinstance(node_entry, dict):
-                for name, info in node_entry.items():
-                    ip = info.get('ip_addr', '')
-                    if ip == 'N/A':
-                        ip = ''
-                    # Normalize device name to consistent capitalization
-                    display_name = normalize_device_name(name)
-                    # Store original name for virsh console (VM names match topo_build.yml)
-                    devices[display_name] = {'ip': ip, 'user_added': False, 'vm_name': name, 'device_category': 'node'}
-
-    # Merge user-added nodes from user_nodes.yaml (for dynamically added nodes)
-    try:
-        if os.path.exists(user_nodes_path):
-            with open(user_nodes_path, 'r') as f:
-                user_data = YAML().load(f)
-            if user_data and 'nodes' in user_data and user_data['nodes']:
-                for node_entry in user_data['nodes']:
-                    if isinstance(node_entry, dict):
-                        for name, info in node_entry.items():
-                            ip = info.get('ip_addr', '')
-                            if ip == 'N/A':
-                                ip = ''
-                            display_name = normalize_device_name(name)
-                            devices[display_name] = {
-                                'ip': ip,
-                                'user_added': True,
-                                'device_type': info.get('device_type', 'other'),
-                                'vm_name': name,
-                                'device_category': 'node'
-                            }
-                pS(f"Merged {len(user_data['nodes'])} user-added nodes into device list")
-    except Exception as e:
-        safe_log('warning', f'Error loading user_nodes.yaml: {e}', event='config', handler='get_all_devices')
-
-    # Merge user-added hosts from user_hosts.yaml (Linux desktop VMs)
-    try:
-        if os.path.exists(user_hosts_path):
-            with open(user_hosts_path, 'r') as f:
-                hosts_data = YAML().load(f)
-            if hosts_data and 'hosts' in hosts_data and hosts_data['hosts']:
-                for host_entry in hosts_data['hosts']:
-                    if isinstance(host_entry, dict):
-                        for name, info in host_entry.items():
-                            ip = info.get('ip_addr', info.get('mgmt_ip', ''))
-                            if ip == 'N/A':
-                                ip = ''
-                            display_name = normalize_device_name(name)
-                            devices[display_name] = {
-                                'ip': ip,
-                                'user_added': True,
-                                'device_type': 'linux_host',
-                                'vm_name': name,
-                                'device_category': 'host',
-                                'supports_novnc': True  # Linux hosts have noVNC desktop access
-                            }
-                pS(f"Merged {len(hosts_data['hosts'])} user-added hosts into device list")
-    except Exception as e:
-        safe_log('warning', f'Error loading user_hosts.yaml: {e}', event='config', handler='get_all_devices')
-
-    # Merge user-added firewalls from user_firewalls.yaml (VyOS firewalls)
-    try:
-        if os.path.exists(user_firewalls_path):
-            with open(user_firewalls_path, 'r') as f:
-                firewalls_data = YAML().load(f)
-            if firewalls_data and 'firewalls' in firewalls_data and firewalls_data['firewalls']:
-                for fw_entry in firewalls_data['firewalls']:
-                    if isinstance(fw_entry, dict):
-                        for name, info in fw_entry.items():
-                            ip = info.get('ip_addr', info.get('mgmt_ip', ''))
-                            if ip == 'N/A':
-                                ip = ''
-                            display_name = normalize_device_name(name)
-                            devices[display_name] = {
-                                'ip': ip,
-                                'user_added': True,
-                                'device_type': 'firewall',
-                                'vm_name': name,
-                                'device_category': 'firewall'
-                            }
-                pS(f"Merged {len(firewalls_data['firewalls'])} user-added firewalls into device list")
-    except Exception as e:
-        safe_log('warning', f'Error loading user_firewalls.yaml: {e}', event='config', handler='get_all_devices')
-
-    # Merge user-added VeloCloud devices from user_velo.yaml
-    try:
-        if os.path.exists(user_velo_path):
-            with open(user_velo_path, 'r') as f:
-                velo_data = YAML().load(f)
-            if velo_data and 'devices' in velo_data and velo_data['devices']:
-                for velo_entry in velo_data['devices']:
-                    if isinstance(velo_entry, dict):
-                        for name, info in velo_entry.items():
-                            ip = info.get('mgmt_ip', '')
-                            if ip == 'N/A':
-                                ip = ''
-                            device_type = info.get('device_type', 'edge')
-                            display_name = normalize_device_name(name)
-                            devices[display_name] = {
-                                'ip': ip,
-                                'user_added': True,
-                                'device_type': f'velo_{device_type}',
-                                'vm_name': name,
-                                'device_category': 'velocloud',
-                                # VCO web UI requires embedded Firefox (proxy abandoned)
-                                # Access via: https://<mgmt_ip>/operator/ with admin@velocloud.local
-                                'supports_webui': False
-                            }
-                pS(f"Merged {len(velo_data['devices'])} user-added VeloCloud devices into device list")
-    except Exception as e:
-        safe_log('warning', f'Error loading user_velo.yaml: {e}', event='config', handler='get_all_devices')
-
-    _ALL_DEVICES_CACHE = devices
-    pS(f"Cached {len(devices)} devices from topo_build.yml + user files")
-    return devices
+# device utility functions (invalidate_devices_cache, get_all_devices, get_device_ip_from_sources)
+# and their cache globals have been moved to handlers/topology_api.py
 
 
 def update_hubspot_handler(email, action, project):
@@ -1151,1807 +802,21 @@ def update_hubspot_handler(email, action, project):
         print(error_msg)
         return {"error": error_msg}
 
-class ToolsHandler(tornado.web.RequestHandler):
-    def post(self):
-        try:
-            # Parse the JSON body of the request
-            data = json.loads(self.request.body)
-            # Extract the three parameters
-            changeLatency = data.get('changeLatency', False)
-            devices = data.get('devices', [])
-            score = data.get('score', 0)
-            # Validate device names to prevent command injection
-            import re
-            for d in devices:
-                if not re.match(r'^[a-zA-Z0-9_-]+$', str(d)):
-                    self.set_status(400)
-                    self.write({"error": f"Invalid device name: {d}"})
-                    return
-            result = subprocess.run(
-                ['please', 'update', 'code',
-                 'ENABLE' if changeLatency else 'DISABLE',
-                 '-d', str(int(score)), '-i', ','.join(devices)],
-                capture_output=True, text=True, timeout=30
-            )
-            
-            # Prepare the response
-            response = {
-                "changeLatency": changeLatency,
-                "devices": devices,
-                "score": score,
-                "result" : result.stdout,
-                "message": "Parameters received successfully"
-            }
-            
-            # Send the response
-            self.set_header("Content-Type", "application/json")
-            self.write(json.dumps(response))
-        
-        except json.JSONDecodeError:
-            safe_log('error', 'Error in ToolsHandler: Invalid JSON in request body', event='error', handler='ToolsHandler')
-            self.set_status(400)
-            self.write({"error": "Invalid JSON in request body"})
-        except ValueError as e:
-            safe_log('error', f'Error in ToolsHandler: {e}', event='error', handler='ToolsHandler')
-            self.set_status(400)
-            self.write({"error": str(e)})
-        except Exception as e:
-            safe_log('error', f'Error in ToolsHandler: {e}', event='error', handler='ToolsHandler')
-            self.set_status(500)
-            self.write({"error": "Internal server error"})
+# ToolsHandler, ViewConfigHandler, BaseUrlHandler, UptimeWithRuntimeHandler,
+# TerminalPageHandler, ConsolePageHandler — moved to handlers/pages.py
 
-class ViewConfigHandler(tornado.web.RequestHandler):
-    def post(self):
-        try:
-            # Parse the JSON body of the request
-            data = json.loads(self.request.body)
 
-            # Extract the parameters
-            devices = data.get('devices', [])
-            # Validate device names to prevent command injection
-            import re
-            for d in devices:
-                if not re.match(r'^[a-zA-Z0-9_-]+$', str(d)):
-                    self.set_status(400)
-                    self.write({"error": f"Invalid device name: {d}"})
-                    return
-            result = subprocess.run(
-                ['sudo', '-S', 'python3', '/home/atdadmin/change-latency.py',
-                 'SHOW', '-i', ','.join(devices)],
-                capture_output=True, text=True, timeout=30
-            )
-            # Prepare the response
-            response = {
-                "devices": devices,
-                "result" : result.stdout,
-                "message": "Parameters received successfully"
-            }
-            
-            # Send the response
-            self.set_header("Content-Type", "application/json")
-            self.write(json.dumps(response))
-        
-        except json.JSONDecodeError:
-            safe_log('error', 'Error in ViewConfigHandler: Invalid JSON in request body', event='error', handler='ViewConfigHandler')
-            self.set_status(400)
-            self.write({"error": "Invalid JSON in request body"})
-        except ValueError as e:
-            safe_log('error', f'Error in ViewConfigHandler: {e}', event='error', handler='ViewConfigHandler')
-            self.set_status(400)
-            self.write({"error": str(e)})
-        except Exception as e:
-            safe_log('error', f'Error in ViewConfigHandler: {e}', event='error', handler='ViewConfigHandler')
-            self.set_status(500)
-            self.write({"error": "Internal server error"})
-class BaseUrlHandler(tornado.web.RequestHandler):
-    def get(self):
-        try:
-            self.set_header("Access-Control-Allow-Origin", "*")
-            with open(ATD_ACCESS_PATH, 'r') as f:
-                host_yaml = YAML().load(f)
-            login_info = host_yaml.get('login_info', {}).get('jump_host', {})
-            response = {
-                "pwd": login_info.get('pw', ''),
-                "user": login_info.get('user', '')
-            }
-            encoded_response = b64encode(json.dumps(response).encode()).decode()
-            self.write({"response": encoded_response})
-        except Exception as e:
-            safe_log('error', f'Error in BaseUrlHandler: {e}', event='error', handler='BaseUrlHandler')
-            self.set_status(500)
-            self.write({"error": str(e)})
 
-class UptimeWithRuntimeHandler(tornado.web.RequestHandler):
-    def get(self):
-        """
-        Handler to provide uptime data with runtime information for timer widget
-        """
-        try:
-            self.set_header("Access-Control-Allow-Origin", "*")
-            self.set_header("Content-Type", "application/json")
-
-            # Get uptime data directly from atd-uptime service
-            try:
-                response = requests.get("http://atd-uptime:50010/uptime", timeout=1)
-                instance_data = response.json()
-
-                # Add runtime from topology metadata
-                if instance_data.get('status') == 'init' and TOPO_DATA and 'labels' in TOPO_DATA and 'runtime' in TOPO_DATA['labels']:
-                    instance_data['runtime'] = int(TOPO_DATA['labels']['runtime'])
-                else:
-                    instance_data['runtime'] = 12
-
-                # Add exam time information if available
-                instance_data['exam_end_time'] = EXAM_END_TIME
-                instance_data['exam_start_time'] = EXAM_START_TIME
-
-                self.write(json.dumps(instance_data))
-            except Exception as e:
-                safe_log('warning', f'Uptime service not ready: {e}', event='uptime', handler='UptimeWithRuntimeHandler')
-                # If uptime service is not ready, return default values
-                self.write(json.dumps({
-                    'boottime': 0,
-                    'uptime': 0,
-                    'runtime': 12,
-                    'status': 'init',
-                    'exam_end_time': EXAM_END_TIME,
-                    'exam_start_time': EXAM_START_TIME
-                }))
-        except Exception as e:
-            safe_log('error', f'Error in UptimeWithRuntimeHandler: {e}', event='error', handler='UptimeWithRuntimeHandler')
-            self.set_status(500)
-            self.write(json.dumps({
-                "error": str(e),
-                "boottime": 0,
-                "uptime": 0,
-                "runtime": 12,
-                "status": "error",
-                "exam_end_time": 0,
-                "exam_start_time": 0
-            }))
-
-class TerminalPageHandler(BaseHandler):
-    """Handler for the tabbed terminal page."""
-
-    def get(self):
-        safe_log('info', 'Terminal page accessed', event='page_view', page='terminal')
-        if not self.current_user:
-            if 'auth' in self.request.arguments:
-                self.redirect('/login?auth={0}'.format(self.get_argument('auth')))
-            else:
-                self.redirect('/login')
-            return
-
-        with open(ATD_ACCESS_PATH, 'r') as f:
-            host_yaml = YAML().load(f)
-        self.render(
-            BASE_PATH + 'terminal.html',
-            topo_title=TITLE,
-            ARISTA_PWD=host_yaml['login_info']['jump_host']['pw'],
-        )
-
-
-class ConsolePageHandler(BaseHandler):
-    """Handler for the serial console page (virsh console access)."""
-
-    def get(self):
-        safe_log('info', 'Console page accessed', event='page_view', page='console')
-        if not self.current_user:
-            if 'auth' in self.request.arguments:
-                self.redirect('/login?auth={0}'.format(self.get_argument('auth')))
-            else:
-                self.redirect('/login')
-            return
-
-        with open(ATD_ACCESS_PATH, 'r') as f:
-            host_yaml = YAML().load(f)
-        self.render(
-            BASE_PATH + 'console.html',
-            topo_title=TITLE,
-            ARISTA_PWD=host_yaml['login_info']['jump_host']['pw'],
-        )
-
-
-class TopologyAPIHandler(BaseHandler):
-    """API endpoint to return topology data for interactive Cytoscape.js diagram."""
-
-    # Thread-safe cache for parsed topology data (30 second TTL)
-    _cache = {}
-    _cache_time = 0
-    _cache_lock = threading.Lock()
-    CACHE_TTL = 30
-    # Track user files modification time for cache invalidation
-    _user_nodes_mtime = 0
-    _user_hosts_mtime = 0
-    _user_firewalls_mtime = 0
-    _user_velo_mtime = 0
-    _user_cloudeos_mtime = 0
-    _user_links_mtime = 0
-
-    @staticmethod
-    def classify_device_type(device_name):
-        """Classify device type based on naming pattern. Uses shared DeviceTypeConfig."""
-        return DeviceTypeConfig.classify_device(device_name)
-
-    @staticmethod
-    def extract_datacenter(device_name):
-        """
-        Extract datacenter identifier from device name.
-        E.g., 'spine1-DC1' -> 'DC1', 'leaf2-DC2' -> 'DC2', 'host1' -> ''
-        Also handles WAN Gateway naming: 'GW11' -> 'DC1', 'GW21' -> 'DC2', 'GW31' -> 'DC3'
-        """
-        import re
-        # Match -DC followed by number or letter at end of name
-        match = re.search(r'-?(DC\d+|dc\d+)$', device_name, re.IGNORECASE)
-        if match:
-            return match.group(1).upper()
-
-        # Handle GW device naming: GW11, GW12 -> DC1, GW21, GW22 -> DC2, GW31 -> DC3
-        # First digit after GW is the DC number
-        if device_name.startswith('GW') and len(device_name) >= 3 and device_name[2].isdigit():
-            dc_num = device_name[2]
-            return f'DC{dc_num}'
-
-        return ''  # No datacenter suffix
-
-    @staticmethod
-    def extract_isp_provider(device_name):
-        """
-        Extract ISP provider identifier from device name.
-        E.g., 'core1-ISP1' -> 'ISP1', 'core2-ISP2' -> 'ISP2', 'internet' -> ''
-        Used for grouping ISP devices by provider in the topology layout.
-        """
-        import re
-        # Match -ISP followed by number
-        match = re.search(r'-?(ISP\d+)$', device_name, re.IGNORECASE)
-        if match:
-            return match.group(1).upper()
-        return ''  # No ISP suffix
-
-    @staticmethod
-    def get_sort_key(device_name):
-        """
-        Generate a sort key for natural ordering of device names.
-        E.g., spine1, spine2, spine10 sorts correctly (not spine1, spine10, spine2)
-        """
-        import re
-        # Split name into text and number parts
-        parts = re.split(r'(\d+)', device_name)
-        result = []
-        for part in parts:
-            if part.isdigit():
-                result.append(int(part))
-            else:
-                result.append(part.lower())
-        return result
-
-    @staticmethod
-    def detect_topology_type(nodes_data):
-        """
-        Detect the type of topology based on device types present.
-        Returns: 'wan' for P-router mesh topologies, 'datacenter' for spine-leaf
-        """
-        device_types = set(node['data']['device_type'] for node in nodes_data)
-
-        has_p_routers = 'p' in device_types
-        has_pe_routers = 'pe' in device_types
-        has_spines = 'spine' in device_types
-        has_leaves = 'leaf' in device_types
-
-        # WAN topology: has P and PE routers, no spines
-        if has_p_routers and has_pe_routers and not has_spines:
-            return 'wan'
-
-        # Datacenter topology: has spines or leaves
-        if has_spines or has_leaves:
-            return 'datacenter'
-
-        # Default to datacenter layout
-        return 'datacenter'
-
-    @staticmethod
-    def extract_site_number(device_name):
-        """
-        Extract site number from device name for WAN topology positioning.
-        E.g., 'PE1' -> 1, 'A2' -> 2, 'PE-1' -> 1, 'SiteA-1' -> 1
-        Returns 1 for "left" side, 2 for "right" side, 0 for center (P routers)
-        """
-        import re
-
-        # Look for trailing number
-        match = re.search(r'[-_]?(\d+)$', device_name)
-        if match:
-            num = int(match.group(1))
-            # Odd numbers = site 1 (left), Even numbers = site 2 (right)
-            return 1 if num % 2 == 1 else 2
-
-        return 0  # No number found - treat as center
-
-    @staticmethod
-    def calculate_wan_positions(nodes_data, edges_data):
-        """
-        Calculate positions for WAN topology with P-router mesh in center.
-        Layout: Left customers -> Left PEs -> P mesh -> Right PEs -> Right customers
-        """
-        import re
-
-        NODE_SPACING_X = 185   # Horizontal spacing
-        NODE_SPACING_Y = 135   # Vertical spacing
-        COLUMN_SPACING = 260   # Extra spacing between columns
-        PADDING = 100
-
-        # Build adjacency for graph analysis
-        adjacency = {}
-        for node in nodes_data:
-            adjacency[node['data']['id']] = set()
-        for edge in edges_data:
-            src = edge['data']['source']
-            tgt = edge['data']['target']
-            if src in adjacency and tgt in adjacency:
-                adjacency[src].add(tgt)
-                adjacency[tgt].add(src)
-
-        # Categorize nodes
-        p_routers = []
-        pe_routers = []
-        customer_devices = []  # CE, hosts, or other edge devices
-        other_devices = []
-
-        for node in nodes_data:
-            dtype = node['data']['device_type']
-            if dtype == 'p':
-                p_routers.append(node)
-            elif dtype == 'pe':
-                pe_routers.append(node)
-            elif DeviceTypeConfig.is_wan_customer_device(dtype) or dtype == 'leaf':
-                # Customer/endpoint devices go to left/right columns in WAN layout
-                # Also includes leafs which often connect to customer equipment
-                customer_devices.append(node)
-            else:
-                other_devices.append(node)
-
-        # Determine which side each PE is on using graph analysis
-        # PEs are on "left" if their connected customers have odd site numbers
-        pe_sides = {}
-        for pe in pe_routers:
-            pe_id = pe['data']['id']
-            pe_neighbors = adjacency.get(pe_id, set())
-
-            # Check connected devices (excluding P routers)
-            side_votes = []
-            for neighbor in pe_neighbors:
-                neighbor_node = next((n for n in nodes_data if n['data']['id'] == neighbor), None)
-                if neighbor_node and neighbor_node['data']['device_type'] != 'p':
-                    site = TopologyAPIHandler.extract_site_number(neighbor)
-                    if site > 0:
-                        side_votes.append(site)
-
-            # Also consider the PE's own name
-            pe_site = TopologyAPIHandler.extract_site_number(pe_id)
-            if pe_site > 0:
-                side_votes.append(pe_site)
-
-            # Majority vote, default to name-based
-            if side_votes:
-                pe_sides[pe_id] = 1 if side_votes.count(1) >= side_votes.count(2) else 2
-            else:
-                pe_sides[pe_id] = pe_site if pe_site > 0 else 1
-
-        # Determine customer sides based on their PE connections
-        customer_sides = {}
-        for cust in customer_devices:
-            cust_id = cust['data']['id']
-            cust_neighbors = adjacency.get(cust_id, set())
-
-            # Find connected PEs
-            for neighbor in cust_neighbors:
-                if neighbor in pe_sides:
-                    customer_sides[cust_id] = pe_sides[neighbor]
-                    break
-
-            # Fallback to name-based
-            if cust_id not in customer_sides:
-                site = TopologyAPIHandler.extract_site_number(cust_id)
-                customer_sides[cust_id] = site if site > 0 else 1
-
-        # Split into columns (left to right)
-        left_customers = [n for n in customer_devices if customer_sides.get(n['data']['id'], 1) == 1]
-        left_pes = [n for n in pe_routers if pe_sides.get(n['data']['id'], 1) == 1]
-        right_pes = [n for n in pe_routers if pe_sides.get(n['data']['id'], 2) == 2]
-        right_customers = [n for n in customer_devices if customer_sides.get(n['data']['id'], 1) == 2]
-
-        # Sort each group by name
-        for group in [left_customers, left_pes, p_routers, right_pes, right_customers]:
-            group.sort(key=lambda n: TopologyAPIHandler.get_sort_key(n['data']['id']))
-
-        # Calculate column X positions
-        columns = [left_customers, left_pes, p_routers, right_pes, right_customers]
-        column_names = ['left_cust', 'left_pe', 'p_mesh', 'right_pe', 'right_cust']
-
-        # Find max height (most nodes in any column)
-        max_height = max(len(col) for col in columns) if columns else 1
-
-        # Position each column
-        current_x = PADDING
-        for col_idx, column in enumerate(columns):
-            if not column:
-                continue
-
-            # Center column vertically
-            col_height = len(column) * NODE_SPACING_Y
-            start_y = PADDING + (max_height * NODE_SPACING_Y - col_height) / 2
-
-            for row_idx, node in enumerate(column):
-                node['position'] = {
-                    'x': current_x,
-                    'y': start_y + row_idx * NODE_SPACING_Y
-                }
-                node['data']['wan_column'] = column_names[col_idx]
-
-            current_x += COLUMN_SPACING
-
-        # Position any other devices at the top
-        if other_devices:
-            other_devices.sort(key=lambda n: TopologyAPIHandler.get_sort_key(n['data']['id']))
-            for idx, node in enumerate(other_devices):
-                node['position'] = {
-                    'x': PADDING + idx * NODE_SPACING_X,
-                    'y': PADDING / 2
-                }
-                node['data']['wan_column'] = 'other'
-
-        return nodes_data
-
-    @staticmethod
-    def calculate_positions(nodes_data, edges_data=None):
-        """
-        Calculate x,y positions for nodes based on topology type.
-        Automatically detects WAN vs datacenter topologies.
-        """
-        import re
-
-        # Detect topology type
-        topo_type = TopologyAPIHandler.detect_topology_type(nodes_data)
-
-        # Use WAN layout for P-router mesh topologies
-        if topo_type == 'wan' and edges_data:
-            return TopologyAPIHandler.calculate_wan_positions(nodes_data, edges_data)
-
-        # Standard datacenter layout (tier-based)
-        # Group nodes by tier, then by grouping key (datacenter or ISP provider)
-        tiers = {}
-        ISP_TIER = DeviceTypeConfig.get_tier('isp')
-
-        for node in nodes_data:
-            device_type = node['data']['device_type']
-            tier = DeviceTypeConfig.get_tier(device_type)
-
-            # For ISP tier, group by ISP provider (ISP1, ISP2) instead of datacenter
-            if tier == ISP_TIER:
-                group_key = TopologyAPIHandler.extract_isp_provider(node['data']['id'])
-            else:
-                group_key = TopologyAPIHandler.extract_datacenter(node['data']['id'])
-
-            if tier not in tiers:
-                tiers[tier] = {}
-            if group_key not in tiers[tier]:
-                tiers[tier][group_key] = []
-            tiers[tier][group_key].append(node)
-
-        # Sort nodes within each tier/group by name (natural sort)
-        for tier in tiers:
-            for group_key in tiers[tier]:
-                tiers[tier][group_key].sort(key=lambda n: TopologyAPIHandler.get_sort_key(n['data']['id']))
-
-        # Calculate positions
-        NODE_SPACING_X = 170   # Horizontal spacing between nodes
-        NODE_SPACING_Y = 185   # Vertical spacing between tiers
-        DC_SPACING = 120       # Extra spacing between datacenter groups
-        PADDING = 100          # Left padding
-
-        # Calculate max width considering groups and spacing
-        # Groups can be DCs (DC1, DC2) or ISP providers (ISP1, ISP2) depending on tier
-        max_width = 0
-        for tier in tiers:
-            tier_width = 0
-            group_keys = sorted(tiers[tier].keys())  # Sort: '', 'DC1', 'DC2' or 'ISP1', 'ISP2'
-            for i, gk in enumerate(group_keys):
-                tier_width += len(tiers[tier][gk]) * NODE_SPACING_X
-                if i < len(group_keys) - 1:  # Add spacing between groups
-                    tier_width += DC_SPACING
-            max_width = max(max_width, tier_width)
-
-        if max_width == 0:
-            max_width = NODE_SPACING_X
-
-        # Position nodes by tier (use row_index to skip empty tiers)
-        row_index = 0
-        for tier_num in sorted(tiers.keys()):
-            tier_groups = tiers[tier_num]
-            group_keys = sorted(tier_groups.keys())  # Sort: '', 'DC1', 'DC2' or 'ISP1', 'ISP2'
-
-            # Calculate this tier's total width
-            tier_width = 0
-            for i, gk in enumerate(group_keys):
-                tier_width += len(tier_groups[gk]) * NODE_SPACING_X
-                if i < len(group_keys) - 1:
-                    tier_width += DC_SPACING
-
-            # Center this tier
-            start_x = PADDING + (max_width - tier_width) / 2
-            current_x = start_x
-
-            for i, group_key in enumerate(group_keys):
-                group_nodes = tier_groups[group_key]
-
-                for node in group_nodes:
-                    node['position'] = {
-                        'x': current_x,
-                        'y': PADDING + row_index * NODE_SPACING_Y
-                    }
-                    # Store grouping info for potential UI use
-                    if tier_num == ISP_TIER:
-                        node['data']['isp_provider'] = group_key if group_key else 'default'
-                        node['data']['datacenter'] = 'shared'  # ISP devices are shared across DCs
-                    else:
-                        node['data']['datacenter'] = group_key if group_key else 'default'
-                    current_x += NODE_SPACING_X
-
-                # Add spacing after each group (except last)
-                if i < len(group_keys) - 1:
-                    current_x += DC_SPACING
-
-            # Increment row index for next tier that has nodes
-            row_index += 1
-
-        return nodes_data
-
-    def parse_topology(self, topo_path):
-        """
-        Parse topo_build.yml and return Cytoscape.js formatted data.
-
-        Returns:
-            dict: Success with 'data' key containing topology
-            dict: Error with 'error' and 'error_type' keys
-        """
-        # Try to open and parse the file
-        try:
-            with open(topo_path, 'r') as f:
-                topo_data = YAML().load(f)
-        except FileNotFoundError:
-            return {'error': f'Topology file not found: {topo_path}', 'error_type': 'not_found'}
-        except PermissionError:
-            return {'error': f'Permission denied accessing: {topo_path}', 'error_type': 'permission'}
-        except Exception as e:
-            safe_log('error', f'Error parsing topology file: {e}', event='config', handler='parse_topology')
-            return {'error': f'Failed to parse topology file: {str(e)}', 'error_type': 'parse_error'}
-
-        # Merge user-added nodes from user_nodes.yaml (for dynamically added nodes)
-        user_nodes_path = '/etc/atd/user_nodes.yaml'
-        try:
-            if os.path.exists(user_nodes_path):
-                with open(user_nodes_path, 'r') as f:
-                    user_data = YAML().load(f)
-                if user_data and 'nodes' in user_data and user_data['nodes']:
-                    # Ensure topo_data has nodes list
-                    if topo_data is None:
-                        topo_data = {'nodes': []}
-                    if 'nodes' not in topo_data:
-                        topo_data['nodes'] = []
-                    # Append user-added nodes
-                    topo_data['nodes'].extend(user_data['nodes'])
-                    pS(f"Merged {len(user_data['nodes'])} user-added nodes from {user_nodes_path}")
-        except Exception as e:
-            safe_log('warning', f'Error loading user_nodes.yaml: {e}', event='config', handler='parse_topology')
-            # Continue without user nodes - don't fail the whole topology load
-
-        # Merge user-added hosts from user_hosts.yaml (Linux desktop VMs)
-        user_hosts_path = '/etc/atd/user_hosts.yaml'
-        try:
-            if os.path.exists(user_hosts_path):
-                with open(user_hosts_path, 'r') as f:
-                    hosts_data = YAML().load(f)
-                if hosts_data and 'hosts' in hosts_data and hosts_data['hosts']:
-                    if topo_data is None:
-                        topo_data = {'nodes': []}
-                    if 'nodes' not in topo_data:
-                        topo_data['nodes'] = []
-                    # Convert hosts to node format for topology
-                    for host_entry in hosts_data['hosts']:
-                        if isinstance(host_entry, dict):
-                            for name, info in host_entry.items():
-                                # Create node entry with linux_host device_type
-                                node_info = {
-                                    'ip_addr': info.get('mgmt_ip', info.get('ip_addr', 'N/A')),
-                                    'device_type': 'linux_host',
-                                    'user_added': True,
-                                    'neighbors': info.get('neighbors', [])
-                                }
-                                topo_data['nodes'].append({name: node_info})
-                    pS(f"Merged {len(hosts_data['hosts'])} user-added hosts from {user_hosts_path}")
-        except Exception as e:
-            safe_log('warning', f'Error loading user_hosts.yaml: {e}', event='config', handler='parse_topology')
-
-        # Merge user-added firewalls from user_firewalls.yaml (VyOS firewalls)
-        user_firewalls_path = '/etc/atd/user_firewalls.yaml'
-        try:
-            if os.path.exists(user_firewalls_path):
-                with open(user_firewalls_path, 'r') as f:
-                    firewalls_data = YAML().load(f)
-                if firewalls_data and 'firewalls' in firewalls_data and firewalls_data['firewalls']:
-                    if topo_data is None:
-                        topo_data = {'nodes': []}
-                    if 'nodes' not in topo_data:
-                        topo_data['nodes'] = []
-                    # Convert firewalls to node format for topology
-                    for fw_entry in firewalls_data['firewalls']:
-                        if isinstance(fw_entry, dict):
-                            for name, info in fw_entry.items():
-                                # Create node entry with firewall device_type
-                                node_info = {
-                                    'ip_addr': info.get('mgmt_ip', info.get('ip_addr', 'N/A')),
-                                    'device_type': 'firewall',
-                                    'user_added': True,
-                                    'neighbors': info.get('neighbors', [])
-                                }
-                                topo_data['nodes'].append({name: node_info})
-                    pS(f"Merged {len(firewalls_data['firewalls'])} user-added firewalls from {user_firewalls_path}")
-        except Exception as e:
-            safe_log('warning', f'Error loading user_firewalls.yaml: {e}', event='config', handler='parse_topology')
-
-        # Merge user-added VeloCloud devices from user_velo.yaml
-        user_velo_path = '/etc/atd/user_velo.yaml'
-        try:
-            if os.path.exists(user_velo_path):
-                with open(user_velo_path, 'r') as f:
-                    velo_data = YAML().load(f)
-                if velo_data and 'devices' in velo_data and velo_data['devices']:
-                    if topo_data is None:
-                        topo_data = {'nodes': []}
-                    if 'nodes' not in topo_data:
-                        topo_data['nodes'] = []
-                    # Convert VeloCloud devices to node format for topology
-                    for velo_entry in velo_data['devices']:
-                        if isinstance(velo_entry, dict):
-                            for name, info in velo_entry.items():
-                                device_type = info.get('device_type', 'edge')
-                                # Create node entry with velo_* device_type
-                                node_info = {
-                                    'ip_addr': info.get('mgmt_ip', 'N/A'),
-                                    'device_type': f'velo_{device_type}',
-                                    'user_added': True,
-                                    'neighbors': info.get('neighbors', [])
-                                }
-                                # Add connections as neighbors for edge drawing
-                                connections = info.get('connections', [])
-                                for conn in connections:
-                                    target = conn.get('target_device', '')
-                                    if target:
-                                        node_info['neighbors'].append({
-                                            'neighborDevice': target,
-                                            'neighborPort': conn.get('target_port', ''),
-                                            'port': conn.get('local_port', '')
-                                        })
-                                topo_data['nodes'].append({name: node_info})
-                    pS(f"Merged {len(velo_data['devices'])} user-added VeloCloud devices from {user_velo_path}")
-        except Exception as e:
-            safe_log('warning', f'Error loading user_velo.yaml: {e}', event='config', handler='parse_topology')
-
-        # Merge user-added CloudEOS devices from user_cloudeos.yaml
-        user_cloudeos_path = '/etc/atd/user_cloudeos.yaml'
-        try:
-            if os.path.exists(user_cloudeos_path):
-                with open(user_cloudeos_path, 'r') as f:
-                    cloudeos_data = YAML().load(f)
-                if cloudeos_data and 'devices' in cloudeos_data and cloudeos_data['devices']:
-                    if topo_data is None:
-                        topo_data = {'nodes': []}
-                    if 'nodes' not in topo_data:
-                        topo_data['nodes'] = []
-                    for device_entry in cloudeos_data['devices']:
-                        if isinstance(device_entry, dict):
-                            for name, info in device_entry.items():
-                                if isinstance(info, dict) and info.get('status') == 'creating':
-                                    continue
-                                node_info = {
-                                    'ip_addr': info.get('ip_addr', 'N/A'),
-                                    'device_type': info.get('device_type', 'other'),
-                                    'device_category': 'cloudeos',
-                                    'user_added': True,
-                                    'neighbors': info.get('neighbors', [])
-                                }
-                                topo_data['nodes'].append({name: node_info})
-                    pS(f"Merged user-added CloudEOS devices from {user_cloudeos_path}")
-        except Exception as e:
-            safe_log('warning', f'Error loading user_cloudeos.yaml: {e}', event='config', handler='parse_topology')
-
-        # Merge user-added links from user_links.yaml
-        # These add neighbor entries to existing topology nodes
-        user_links_path = '/etc/atd/user_links.yaml'
-        try:
-            if os.path.exists(user_links_path):
-                with open(user_links_path, 'r') as f:
-                    links_data = YAML().load(f)
-                if links_data and 'links' in links_data and links_data['links']:
-                    links_merged = 0
-                    for link in links_data['links']:
-                        source = link.get('source_device', '')
-                        source_port = link.get('source_port', '')
-                        target = link.get('target_device', '')
-                        target_port = link.get('target_port', '')
-                        if source and target:
-                            # Add neighbor entry to source node
-                            for node_entry in topo_data['nodes']:
-                                if isinstance(node_entry, dict):
-                                    for node_name in node_entry:
-                                        if node_name.lower() == source.lower():
-                                            neighbors = node_entry[node_name].setdefault('neighbors', [])
-                                            neighbors.append({
-                                                'neighborDevice': target,
-                                                'neighborPort': target_port,
-                                                'port': source_port,
-                                                'user_added': True
-                                            })
-                                            links_merged += 1
-                    if links_merged > 0:
-                        pS(f"Merged {links_merged} user-added links from {user_links_path}")
-        except Exception as e:
-            safe_log('warning', f'Error loading user_links.yaml: {e}', event='config', handler='parse_topology')
-
-        # Validate topo_data structure
-        if topo_data is None:
-            return {'error': 'Topology file is empty', 'error_type': 'empty_file'}
-
-        if 'nodes' not in topo_data:
-            return {'error': 'Topology file missing "nodes" key', 'error_type': 'invalid_format'}
-
-        if not topo_data['nodes']:
-            # Empty nodes list - return empty topology (valid case)
-            pS("Warning: Topology file has no nodes")
-            return {
-                'data': {
-                    'metadata': {
-                        'topology_name': TOPO,
-                        'eos_type': EOS_TYPE,
-                        'node_count': 0,
-                        'edge_count': 0,
-                        'generated_at': datetime.now().isoformat()
-                    },
-                    'nodes': [],
-                    'edges': []
-                }
-            }
-
-        nodes = []
-        edges = []
-        edge_set = set()  # Track edges to avoid duplicates
-
-        # First pass: collect all valid node names and build normalization mapping
-        # Maps raw_name (from YAML) -> normalized_name (for display and API consistency)
-        valid_node_names = set()
-        name_mapping = {}  # raw_name -> normalized_name
-        for node_entry in topo_data['nodes']:
-            if isinstance(node_entry, dict):
-                for raw_name in node_entry.keys():
-                    normalized = normalize_device_name(raw_name)
-                    valid_node_names.add(raw_name)
-                    name_mapping[raw_name] = normalized
-
-        # Second pass: build nodes and edges
-        for node_entry in topo_data['nodes']:
-            # Validate node entry is a dict
-            if not isinstance(node_entry, dict):
-                pS(f"Warning: Invalid node entry format (not a dict): {node_entry}")
-                continue
-
-            # Each entry is a dict with device name as key
-            for raw_device_name, device_info in node_entry.items():
-                # Get normalized display name for API consistency
-                display_name = name_mapping.get(raw_device_name, raw_device_name)
-
-                # Validate device_info is a dict
-                if not isinstance(device_info, dict):
-                    pS(f"Warning: Invalid device info for {raw_device_name} (not a dict)")
-                    continue
-
-                # Use explicit device_type if provided (for user-added nodes),
-                # otherwise classify from device name
-                device_type = device_info.get('device_type') or self.classify_device_type(raw_device_name)
-                ip_addr = device_info.get('ip_addr', 'N/A')
-                sys_mac = device_info.get('sys_mac', 'N/A')
-                neighbors = device_info.get('neighbors', [])
-                # Check if this is a user-added node (from user_nodes.yaml)
-                user_added = device_info.get('user_added', False)
-
-                # Validate neighbors is a list
-                if not isinstance(neighbors, list):
-                    pS(f"Warning: Invalid neighbors format for {raw_device_name}")
-                    neighbors = []
-
-                # Build port info for tooltip
-                ports = []
-                for neighbor in neighbors:
-                    if not isinstance(neighbor, dict):
-                        continue
-
-                    neighbor_device_raw = neighbor.get('neighborDevice', '')
-                    # Get normalized neighbor name for display
-                    neighbor_device_display = name_mapping.get(neighbor_device_raw, neighbor_device_raw)
-
-                    ports.append({
-                        'port': neighbor.get('port', ''),
-                        'neighbor': neighbor_device_display,  # Use normalized name for display
-                        'neighbor_port': neighbor.get('neighborPort', '')
-                    })
-
-                    # Create edge only if both nodes exist (prevents Cytoscape.js errors)
-                    if neighbor_device_raw and neighbor_device_raw in valid_node_names:
-                        # Get port values with None-safety
-                        device_port = neighbor.get('port') or ''
-                        neighbor_port = neighbor.get('neighborPort') or ''
-
-                        # Create edge key that includes ports to support multiple links
-                        # between the same device pair (e.g., MLAG, port-channel, redundancy)
-                        # Use normalized names for edge keys
-                        port_pair = tuple(sorted([device_port, neighbor_port]))
-                        edge_key = (tuple(sorted([display_name, neighbor_device_display])), port_pair)
-
-                        if edge_key not in edge_set:
-                            edge_set.add(edge_key)
-
-                            # Use alphabetically sorted order for source/target to ensure
-                            # consistent port assignment regardless of processing order.
-                            # edge_key[0][0] is the alphabetically first device name.
-                            sorted_devices = edge_key[0]
-                            if display_name == sorted_devices[0]:
-                                # Current device is alphabetically first, so it's the source.
-                                # Ports stay as-is: device_port -> source, neighbor_port -> target
-                                source_node = display_name
-                                target_node = neighbor_device_display
-                                source_port = device_port
-                                target_port = neighbor_port
-                            else:
-                                # Neighbor device is alphabetically first, so it becomes source.
-                                # Since we're processing from device's perspective, swap ports:
-                                # neighbor_port belongs to the alphabetically-first (source) node
-                                # device_port belongs to the alphabetically-second (target) node
-                                source_node = neighbor_device_display
-                                target_node = display_name
-                                source_port = neighbor_port
-                                target_port = device_port
-
-                            # Use unique edge ID that includes ports to support parallel links
-                            edge_id = f"{source_node}-{target_node}-{source_port}-{target_port}"
-                            edge_data = {
-                                'id': edge_id,
-                                'source': source_node,
-                                'target': target_node,
-                                'source_port': source_port,
-                                'target_port': target_port
-                            }
-                            # Mark user-added links so the UI can show Remove Link option
-                            # and style them differently on the diagram
-                            edge_entry = {'data': edge_data}
-                            if neighbor.get('user_added'):
-                                edge_data['user_added'] = True
-                                edge_entry['classes'] = 'edge-user-added'
-                            edges.append(edge_entry)
-                    elif neighbor_device_raw:
-                        pS(f"Warning: Skipping edge {display_name}->{neighbor_device_display}: target node not in topology")
-
-                # Create node with normalized display name as ID
-                # Keep vm_name for virsh console access (uses original name from YAML)
-                # Include device_category so the UI can route delete requests correctly
-                device_category = device_info.get('device_category', 'node')
-                nodes.append({
-                    'data': {
-                        'id': display_name,
-                        'label': display_name,
-                        'ip': ip_addr,
-                        'sys_mac': sys_mac,
-                        'device_type': device_type,
-                        'device_category': device_category,
-                        'status': 'unknown',
-                        'ports': ports,
-                        'user_added': user_added,
-                        'vm_name': raw_device_name  # Original name for virsh console
-                    },
-                    'classes': f"device-type-{device_type} status-unknown"
-                })
-
-        # Calculate positions based on device type tiers and natural name sorting
-        # Pass edges for WAN topology detection which uses graph adjacency
-        nodes = self.calculate_positions(nodes, edges)
-
-        return {
-            'data': {
-                'metadata': {
-                    'topology_name': TOPO,
-                    'eos_type': EOS_TYPE,
-                    'node_count': len(nodes),
-                    'edge_count': len(edges),
-                    'generated_at': datetime.now().isoformat()
-                },
-                'nodes': nodes,
-                'edges': edges
-            }
-        }
-
-    def options(self):
-        """Handle CORS preflight requests."""
-        self.set_header("Access-Control-Allow-Origin", "*")
-        self.set_header("Access-Control-Allow-Methods", "GET, OPTIONS")
-        self.set_header("Access-Control-Allow-Headers", "Content-Type, Authorization")
-        self.set_header("Access-Control-Max-Age", "86400")  # Cache preflight for 24 hours
-        self.set_status(204)
-        self.finish()
-
-    def get(self):
-        if not self.current_user:
-            self.set_status(401)
-            self.write(json.dumps({'error': 'Authentication required'}))
-            return
-
-        safe_log('info', 'Topology API requested', event='api', endpoint='topology')
-        self.set_header("Content-Type", "application/json")
-        self.set_header("Access-Control-Allow-Origin", "*")
-
-        try:
-            current_time = time.time()
-
-            # Check if any user file was modified (invalidates cache)
-            user_nodes_path = '/etc/atd/user_nodes.yaml'
-            user_hosts_path = '/etc/atd/user_hosts.yaml'
-            user_firewalls_path = '/etc/atd/user_firewalls.yaml'
-            user_velo_path = '/etc/atd/user_velo.yaml'
-            user_cloudeos_path = '/etc/atd/user_cloudeos.yaml'
-            user_links_path = '/etc/atd/user_links.yaml'
-
-            user_nodes_mtime = os.path.getmtime(user_nodes_path) if os.path.exists(user_nodes_path) else 0
-            user_hosts_mtime = os.path.getmtime(user_hosts_path) if os.path.exists(user_hosts_path) else 0
-            user_firewalls_mtime = os.path.getmtime(user_firewalls_path) if os.path.exists(user_firewalls_path) else 0
-            user_velo_mtime = os.path.getmtime(user_velo_path) if os.path.exists(user_velo_path) else 0
-            user_cloudeos_mtime = os.path.getmtime(user_cloudeos_path) if os.path.exists(user_cloudeos_path) else 0
-            user_links_mtime = os.path.getmtime(user_links_path) if os.path.exists(user_links_path) else 0
-
-            # Thread-safe cache check - invalidate if any user file changed
-            with TopologyAPIHandler._cache_lock:
-                cache_valid = (
-                    TopologyAPIHandler._cache and
-                    current_time - TopologyAPIHandler._cache_time < TopologyAPIHandler.CACHE_TTL and
-                    user_nodes_mtime <= TopologyAPIHandler._user_nodes_mtime and
-                    user_hosts_mtime <= TopologyAPIHandler._user_hosts_mtime and
-                    user_firewalls_mtime <= TopologyAPIHandler._user_firewalls_mtime and
-                    user_velo_mtime <= TopologyAPIHandler._user_velo_mtime and
-                    user_cloudeos_mtime <= TopologyAPIHandler._user_cloudeos_mtime and
-                    user_links_mtime <= TopologyAPIHandler._user_links_mtime
-                )
-                if cache_valid:
-                    self.write(json.dumps(TopologyAPIHandler._cache))
-                    return
-
-            # Parse topology file (outside lock to avoid blocking)
-            topo_path = f"/opt/atd/topologies/{TOPO}/topo_build.yml"
-            result = self.parse_topology(topo_path)
-
-            # Check for errors
-            if 'error' in result:
-                error_type = result.get('error_type', 'unknown')
-                if error_type == 'not_found':
-                    self.set_status(404)
-                elif error_type == 'permission':
-                    self.set_status(403)
-                else:
-                    self.set_status(500)
-                self.write(json.dumps({'error': result['error']}))
-                return
-
-            topology_data = result['data']
-
-            # Thread-safe cache update (include user file mtimes for invalidation)
-            with TopologyAPIHandler._cache_lock:
-                TopologyAPIHandler._cache = topology_data
-                TopologyAPIHandler._cache_time = current_time
-                TopologyAPIHandler._user_nodes_mtime = user_nodes_mtime
-                TopologyAPIHandler._user_hosts_mtime = user_hosts_mtime
-                TopologyAPIHandler._user_firewalls_mtime = user_firewalls_mtime
-                TopologyAPIHandler._user_velo_mtime = user_velo_mtime
-                TopologyAPIHandler._user_cloudeos_mtime = user_cloudeos_mtime
-                TopologyAPIHandler._user_links_mtime = user_links_mtime
-
-            self.write(json.dumps(topology_data))
-
-        except Exception as e:
-            safe_log('error', f'Error in TopologyAPIHandler: {e}', event='error', handler='TopologyAPIHandler')
-            traceback.print_exc()
-            self.set_status(500)
-            self.write(json.dumps({'error': f'Internal server error: {str(e)}'}))
-
-
-class DevicesAPIHandler(BaseHandler):
-    """API endpoint to return device list grouped by type for terminal page."""
-
-    def get(self):
-        if not self.current_user:
-            self.set_status(401)
-            self.write(json.dumps({'error': 'Authentication required'}))
-            return
-
-        safe_log('info', 'Devices API requested', event='api', endpoint='devices')
-        self.set_header("Content-Type", "application/json")
-        self.set_header("Access-Control-Allow-Origin", "*")
-
-        try:
-            # Get devices from topo_build.yml, user_nodes.yaml, user_hosts.yaml, user_firewalls.yaml
-            nodes = get_all_devices()
-
-            # Group devices using shared DeviceTypeConfig
-            # User-added items go to separate groups by category at the end
-            groups = {}
-            user_nodes_group = []
-            user_hosts_group = []
-            user_firewalls_group = []
-            user_velocloud_group = []
-
-            for device_name, device_info in nodes.items():
-                is_user_added = device_info.get('user_added', False)
-                device_category = device_info.get('device_category', 'node')
-
-                # Build device entry with new flags
-                # Console only supported for KVM labs (virsh console), not cEOS
-                supports_console = EOS_TYPE != 'container-labs'
-                # Include original VM name for virsh console connections
-                vm_name = device_info.get('vm_name', device_name)
-                # Linux hosts support noVNC desktop access
-                supports_novnc = device_info.get('supports_novnc', False)
-                # VeloCloud Orchestrator supports web UI access
-                supports_webui = device_info.get('supports_webui', False)
-
-                device_entry = {
-                    'name': device_name,
-                    'vmName': vm_name,  # Original name for virsh console
-                    'ip': device_info.get('ip', ''),
-                    'userAdded': is_user_added,
-                    'supportsConsole': supports_console,
-                    'supportsNoVnc': supports_novnc,
-                    'supportsWebUI': supports_webui,
-                }
-
-                if is_user_added:
-                    # User-added devices go to category-specific groups
-                    if device_category == 'host':
-                        user_hosts_group.append(device_entry)
-                    elif device_category == 'firewall':
-                        user_firewalls_group.append(device_entry)
-                    elif device_category == 'velocloud':
-                        user_velocloud_group.append(device_entry)
-                    else:
-                        user_nodes_group.append(device_entry)
-                else:
-                    # Regular nodes grouped by device type
-                    device_type = device_info.get('device_type') or DeviceTypeConfig.classify_device(device_name)
-                    group_name = DeviceTypeConfig.get_group_name(device_type)
-
-                    if group_name not in groups:
-                        groups[group_name] = []
-                    groups[group_name].append(device_entry)
-
-            # Sort devices within each group and format result
-            # Order groups by tier (using first device type that maps to each group)
-            group_order = DeviceTypeConfig.get_all_group_names()
-
-            result = []
-            for group_name in group_order:
-                if group_name in groups and groups[group_name]:
-                    devices = sorted(groups[group_name], key=lambda x: x['name'])
-                    result.append({
-                        'group': group_name,
-                        'devices': devices
-                    })
-
-            # Add any remaining groups not in the predefined order
-            for group_name in sorted(groups.keys()):
-                if group_name not in group_order and groups[group_name]:
-                    devices = sorted(groups[group_name], key=lambda x: x['name'])
-                    result.append({
-                        'group': group_name,
-                        'devices': devices
-                    })
-
-            # Add User Nodes group if there are any
-            if user_nodes_group:
-                result.append({
-                    'group': 'User Nodes',
-                    'devices': sorted(user_nodes_group, key=lambda x: x['name'])
-                })
-
-            # Add User Hosts group if there are any (Linux desktop VMs)
-            if user_hosts_group:
-                result.append({
-                    'group': 'Linux Hosts',
-                    'devices': sorted(user_hosts_group, key=lambda x: x['name'])
-                })
-
-            # Add User Firewalls group if there are any (VyOS firewalls)
-            if user_firewalls_group:
-                result.append({
-                    'group': 'Firewalls',
-                    'devices': sorted(user_firewalls_group, key=lambda x: x['name'])
-                })
-
-            # Add VeloCloud group if there are any (VeloCloud Edge/Gateway/Orchestrator)
-            if user_velocloud_group:
-                result.append({
-                    'group': 'VeloCloud',
-                    'devices': sorted(user_velocloud_group, key=lambda x: x['name'])
-                })
-
-            self.write(json.dumps({
-                'topology': TITLE,
-                'eosType': EOS_TYPE,
-                'groups': result
-            }))
-
-        except FileNotFoundError as e:
-            safe_log('error', f'Error in DevicesAPIHandler: file not found: {e}', event='error', handler='DevicesAPIHandler')
-            self.set_status(503)
-            self.write(json.dumps({
-                'error': 'Device configuration not available',
-                'detail': 'The topology configuration file could not be found. Please wait for the lab to finish initializing.',
-                'retry': True
-            }))
-
-        except json.JSONDecodeError as e:
-            safe_log('error', f'Error in DevicesAPIHandler: parse error: {e}', event='error', handler='DevicesAPIHandler')
-            traceback.print_exc()
-            self.set_status(500)
-            self.write(json.dumps({
-                'error': 'Configuration error',
-                'detail': 'The device configuration could not be parsed.',
-                'retry': False
-            }))
-
-        except Exception as e:
-            safe_log('error', f'Error in DevicesAPIHandler: {e}', event='error', handler='DevicesAPIHandler')
-            traceback.print_exc()
-            self.set_status(500)
-            self.write(json.dumps({
-                'error': 'Internal server error',
-                'detail': str(e),
-                'retry': True
-            }))
-
-
-class DeviceTypesAPIHandler(BaseHandler):
-    """API endpoint to return device type metadata for frontend."""
-
-    def get(self):
-        if not self.current_user:
-            self.set_status(401)
-            self.write(json.dumps({'error': 'Authentication required'}))
-            return
-
-        self.set_header("Content-Type", "application/json")
-        self.set_header("Access-Control-Allow-Origin", "*")
-
-        try:
-            metadata = DeviceTypeConfig.export_for_frontend()
-            self.write(json.dumps(metadata))
-        except Exception as e:
-            safe_log('error', f'Error in DeviceTypesAPIHandler: {e}', event='error', handler='DeviceTypesAPIHandler')
-            self.set_status(500)
-            self.write(json.dumps({'error': str(e)}))
-
-
-class InterfaceStatsAPIHandler(BaseHandler):
-    """API endpoint for interface statistics via eAPI."""
-
-    # Cache: {device_interface: (timestamp, data)}
-    _cache = {}
-    _cache_lock = threading.Lock()
-    CACHE_TTL = 10  # seconds
-
-    # Rate calculation: store previous readings for rate computation
-    _previous_counters = {}
-    _previous_lock = threading.Lock()
-
-    def get(self):
-        if not self.current_user:
-            self.set_status(401)
-            self.write(json.dumps({'error': 'Authentication required'}))
-            return
-
-        safe_log('info', 'Interface stats requested', event='api', endpoint='interface_stats', device=str(self.get_argument('device', 'unknown')))
-        self.set_header("Content-Type", "application/json")
-        self.set_header("Access-Control-Allow-Origin", "*")
-
-        device = self.get_argument('device', None)
-        interface = self.get_argument('interface', None)
-
-        if not device or not interface:
-            self.set_status(400)
-            self.write(json.dumps({'error': 'device and interface parameters required'}))
-            return
-
-        try:
-            stats = self.get_interface_stats(device, interface)
-            self.write(json.dumps(stats))
-        except Exception as e:
-            error_str = str(e)
-            # Check for authentication failures - device is up but not configured
-            if 'Unauthorized' in error_str or 'Bad username' in error_str or 'authentication' in error_str.lower():
-                safe_log('warning', f'InterfaceStatsAPIHandler: Auth failed for {device}', event='api', handler='InterfaceStatsAPIHandler', device=str(device))
-                self.write(json.dumps({
-                    'device': device,
-                    'interface': interface,
-                    'status': 'unconfigured',
-                    'error': 'Device reachable but authentication failed (not yet configured)'
-                }))
-            elif 'Interface does not exist' in error_str or 'Invalid input' in error_str:
-                self.set_status(404)
-                self.write(json.dumps({
-                    'device': device,
-                    'interface': interface,
-                    'status': 'not_found',
-                    'error': f'Interface {interface} does not exist on {device}'
-                }))
-            elif 'timed out' in error_str.lower() or 'connection timed out' in error_str.lower():
-                self.write(json.dumps({
-                    'device': device,
-                    'interface': interface,
-                    'status': 'down',
-                    'error': f'Device {device} is unreachable'
-                }))
-            else:
-                safe_log('error', f'InterfaceStatsAPIHandler error: {e}', event='api', handler='InterfaceStatsAPIHandler')
-                self.set_status(500)
-                self.write(json.dumps({'error': error_str}))
-
-    def get_interface_stats(self, device_name, interface_name):
-        """Query EOS device for interface counters via eAPI."""
-        cache_key = f"{device_name}:{interface_name}"
-        current_time = time.time()
-
-        # Check cache
-        with self._cache_lock:
-            if cache_key in self._cache:
-                timestamp, data = self._cache[cache_key]
-                if current_time - timestamp < self.CACHE_TTL:
-                    return data
-
-        # Get device IP from topology
-        device_ip = get_device_ip_from_sources(device_name)
-        if not device_ip:
-            raise ValueError(f"Device {device_name} not found in topology")
-
-        # Get credentials from ACCESS_INFO
-        with open(ATD_ACCESS_PATH, 'r') as f:
-            host_yaml = YAML().load(f)
-        username = host_yaml['login_info']['jump_host']['user']
-        password = host_yaml['login_info']['jump_host']['pw']
-
-        # Connect via eAPI
-        try:
-            connection = pyeapi.connect(
-                host=device_ip,
-                username=username,
-                password=password,
-                transport='https',
-                timeout=10
-            )
-
-            # Execute show interfaces command
-            result = connection.execute([f"show interfaces {interface_name}"])
-
-            # Parse interface data
-            interfaces = result.get('result', [{}])[0].get('interfaces', {})
-            intf_data = interfaces.get(interface_name, {})
-
-            if not intf_data:
-                raise ValueError(f"Interface {interface_name} not found on {device_name}")
-
-            counters = intf_data.get('interfaceCounters', {})
-            bandwidth = intf_data.get('bandwidth', 0)
-
-            # Calculate rates from counter deltas
-            rates = self.calculate_rates(cache_key, counters, current_time)
-
-            # Calculate utilization percentage
-            if bandwidth > 0:
-                utilization_in = (rates['in_rate_bps'] / bandwidth) * 100
-                utilization_out = (rates['out_rate_bps'] / bandwidth) * 100
-            else:
-                utilization_in = 0
-                utilization_out = 0
-
-            stats = {
-                'device': device_name,
-                'interface': interface_name,
-                'stats': {
-                    'in_octets': counters.get('inOctets', 0),
-                    'out_octets': counters.get('outOctets', 0),
-                    'in_rate_bps': rates['in_rate_bps'],
-                    'out_rate_bps': rates['out_rate_bps'],
-                    'in_packets': counters.get('inUcastPkts', 0) + counters.get('inMulticastPkts', 0) + counters.get('inBroadcastPkts', 0),
-                    'out_packets': counters.get('outUcastPkts', 0) + counters.get('outMulticastPkts', 0) + counters.get('outBroadcastPkts', 0),
-                    'in_errors': counters.get('inErrors', 0) + counters.get('inputErrorsDetail', {}).get('crcErrors', 0),
-                    'out_errors': counters.get('outErrors', 0),
-                    'in_discards': counters.get('inDiscards', 0),
-                    'out_discards': counters.get('outDiscards', 0),
-                    'speed_bps': bandwidth,
-                    'utilization_in': round(utilization_in, 2),
-                    'utilization_out': round(utilization_out, 2),
-                    'operational_status': intf_data.get('interfaceStatus', 'unknown'),
-                    'line_protocol': intf_data.get('lineProtocolStatus', 'unknown'),
-                    'description': intf_data.get('description', ''),
-                    'last_updated': datetime.now().isoformat()
-                }
-            }
-
-            # Update cache
-            with self._cache_lock:
-                self._cache[cache_key] = (current_time, stats)
-
-            return stats
-
-        except (socket.timeout, OSError) as e:
-            raise ValueError(f"Connection timed out to {device_name} ({device_ip})")
-        except pyeapi.eapilib.ConnectionError as e:
-            # Preserve auth failure info for upstream handler to detect 'unconfigured' status
-            error_str = str(e)
-            if 'Unauthorized' in error_str or 'Bad username' in error_str:
-                raise ValueError(f"Unauthorized: Cannot authenticate to {device_name} ({device_ip}): {e}")
-            raise ValueError(f"Cannot connect to {device_name} ({device_ip}): {e}")
-        except pyeapi.eapilib.CommandError as e:
-            raise ValueError(f"Command error on {device_name}: {e}")
-
-    def calculate_rates(self, cache_key, current_counters, current_time):
-        """Calculate bit rates from counter deltas."""
-        with self._previous_lock:
-            if cache_key in self._previous_counters:
-                prev_time, prev_counters = self._previous_counters[cache_key]
-                time_delta = current_time - prev_time
-
-                if time_delta > 0:
-                    in_octet_delta = current_counters.get('inOctets', 0) - prev_counters.get('inOctets', 0)
-                    out_octet_delta = current_counters.get('outOctets', 0) - prev_counters.get('outOctets', 0)
-
-                    # Handle counter wrap (unlikely but possible)
-                    if in_octet_delta < 0:
-                        in_octet_delta = current_counters.get('inOctets', 0)
-                    if out_octet_delta < 0:
-                        out_octet_delta = current_counters.get('outOctets', 0)
-
-                    in_rate = (in_octet_delta * 8) / time_delta
-                    out_rate = (out_octet_delta * 8) / time_delta
-                else:
-                    in_rate = out_rate = 0
-            else:
-                # First reading, no rate available yet
-                in_rate = out_rate = 0
-
-            # Store current reading for next calculation
-            self._previous_counters[cache_key] = (current_time, {
-                'inOctets': current_counters.get('inOctets', 0),
-                'outOctets': current_counters.get('outOctets', 0)
-            })
-
-            return {'in_rate_bps': round(in_rate, 2), 'out_rate_bps': round(out_rate, 2)}
-
-
-class DeviceStatusAPIHandler(BaseHandler):
-    """API endpoint to check device reachability via eAPI."""
-
-    # Cache: {device: (timestamp, status)}
-    _cache = {}
-    _cache_lock = threading.Lock()
-    CACHE_TTL = 30  # seconds - longer cache for status checks
-
-    def get(self):
-        if not self.current_user:
-            self.set_status(401)
-            self.write(json.dumps({'error': 'Authentication required'}))
-            return
-
-        self.set_header("Content-Type", "application/json")
-        self.set_header("Access-Control-Allow-Origin", "*")
-
-        device = self.get_argument('device', None)
-
-        # If no device specified, check all devices
-        if device:
-            try:
-                status = self.check_device_status(device)
-                self.write(json.dumps(status))
-            except Exception as e:
-                self.write(json.dumps({
-                    'device': device,
-                    'status': 'error',
-                    'error': str(e)
-                }))
-        else:
-            # Check all devices in topology
-            statuses = self.check_all_devices()
-            self.write(json.dumps({'devices': statuses}))
-
-    def check_device_status(self, device_name):
-        """Check if a single device is reachable. Uses eAPI for EOS devices, ping for hosts/firewalls."""
-        cache_key = device_name
-        current_time = time.time()
-
-        # Check cache
-        with self._cache_lock:
-            if cache_key in self._cache:
-                timestamp, data = self._cache[cache_key]
-                if current_time - timestamp < self.CACHE_TTL:
-                    return data
-
-        # Get device info from topology (includes device_category for hosts/firewalls)
-        all_devices = get_all_devices()
-        device_info = all_devices.get(device_name, {})
-        device_ip = device_info.get('ip', '')
-        device_category = device_info.get('device_category', 'node')
-
-        # Fallback to old method if not found in get_all_devices
-        if not device_ip:
-            device_ip = get_device_ip_from_sources(device_name)
-
-        safe_log('info', f'Device status check: {device_name} -> {device_ip}', event='api', handler='DeviceStatusAPIHandler', device=str(device_name))
-        if not device_ip:
-            result = {
-                'device': device_name,
-                'status': 'unknown',
-                'error': 'Device not found in topology'
-            }
-            return result
-
-        # For hosts and firewalls, use ping instead of eAPI
-        if device_category in ('host', 'firewall'):
-            result = self._check_device_via_ping(device_name, device_ip, device_category)
-        else:
-            # For EOS devices, use eAPI
-            result = self._check_device_via_eapi(device_name, device_ip)
-
-        # Update cache
-        with self._cache_lock:
-            self._cache[cache_key] = (current_time, result)
-
-        return result
-
-    def _check_device_via_ping(self, device_name, device_ip, device_category):
-        """Check if a host or firewall is reachable via ping or TCP check."""
-        import subprocess
-        import socket
-
-        device_type_label = 'Linux Host' if device_category == 'host' else 'VyOS Firewall'
-
-        # Try ping first
-        try:
-            # Quick ping with 1 second timeout
-            result = subprocess.run(
-                ['ping', '-c', '1', '-W', '1', device_ip],
-                capture_output=True,
-                timeout=3
-            )
-
-            if result.returncode == 0:
-                return {
-                    'device': device_name,
-                    'ip': device_ip,
-                    'status': 'up',
-                    'version': device_type_label,
-                    'last_check': datetime.now().isoformat()
-                }
-            else:
-                return {
-                    'device': device_name,
-                    'ip': device_ip,
-                    'status': 'down',
-                    'error': 'Ping failed',
-                    'last_check': datetime.now().isoformat()
-                }
-        except subprocess.TimeoutExpired:
-            return {
-                'device': device_name,
-                'ip': device_ip,
-                'status': 'down',
-                'error': 'Ping timeout',
-                'last_check': datetime.now().isoformat()
-            }
-        except FileNotFoundError:
-            # ping command not available, fallback to TCP check on port 22 (SSH)
-            safe_log('warning', f'Ping unavailable for {device_name}, trying TCP', event='api', handler='DeviceStatusAPIHandler', device=str(device_name))
-            pass
-        except Exception as e:
-            # Log the error but try TCP fallback
-            safe_log('warning', f'Ping failed for {device_name}: {e}, trying TCP', event='api', handler='DeviceStatusAPIHandler', device=str(device_name))
-            pass
-
-        # Fallback: TCP check on SSH port (22)
-        try:
-            sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            sock.settimeout(2)
-            result = sock.connect_ex((device_ip, 22))
-            sock.close()
-
-            if result == 0:
-                return {
-                    'device': device_name,
-                    'ip': device_ip,
-                    'status': 'up',
-                    'version': device_type_label,
-                    'last_check': datetime.now().isoformat()
-                }
-            else:
-                return {
-                    'device': device_name,
-                    'ip': device_ip,
-                    'status': 'down',
-                    'error': f'TCP port 22 not responding (code {result})',
-                    'last_check': datetime.now().isoformat()
-                }
-        except Exception as e:
-            safe_log('error', f'TCP check failed for {device_name}: {e}', event='api', handler='DeviceStatusAPIHandler', device=str(device_name))
-            return {
-                'device': device_name,
-                'ip': device_ip,
-                'status': 'down',
-                'error': f'Unreachable: {str(e)}',
-                'last_check': datetime.now().isoformat()
-            }
-
-    def _check_device_via_eapi(self, device_name, device_ip):
-        """Check if an EOS device is reachable via eAPI."""
-        # Get credentials from ACCESS_INFO
-        try:
-            with open(ATD_ACCESS_PATH, 'r') as f:
-                host_yaml = YAML().load(f)
-            username = host_yaml['login_info']['jump_host']['user']
-            password = host_yaml['login_info']['jump_host']['pw']
-        except Exception as e:
-            return {
-                'device': device_name,
-                'status': 'error',
-                'error': f'Cannot read credentials: {e}'
-            }
-
-        # Try to connect via eAPI
-        try:
-            connection = pyeapi.connect(
-                host=device_ip,
-                username=username,
-                password=password,
-                transport='https',
-                timeout=5  # Short timeout for status check
-            )
-
-            # Simple command to verify connectivity
-            result_cmd = connection.execute(['show version'])
-            version = result_cmd.get('result', [{}])[0].get('version', 'unknown')
-
-            return {
-                'device': device_name,
-                'ip': device_ip,
-                'status': 'up',
-                'version': version,
-                'last_check': datetime.now().isoformat()
-            }
-
-        except (socket.timeout, OSError) as e:
-            # Network-level failures: timeout, connection refused, unreachable
-            return {
-                'device': device_name,
-                'ip': device_ip,
-                'status': 'down',
-                'error': f'Connection timed out ({device_ip})',
-                'last_check': datetime.now().isoformat()
-            }
-        except pyeapi.eapilib.ConnectionError as e:
-            # pyeapi raises ConnectionError for auth failures with "Unauthorized" message
-            error_str = str(e)
-            if 'Unauthorized' in error_str or 'Bad username' in error_str or 'authentication' in error_str.lower():
-                return {
-                    'device': device_name,
-                    'ip': device_ip,
-                    'status': 'unconfigured',
-                    'error': 'Device reachable but authentication failed (not yet configured)',
-                    'last_check': datetime.now().isoformat()
-                }
-            return {
-                'device': device_name,
-                'ip': device_ip,
-                'status': 'down',
-                'error': 'Connection failed',
-                'last_check': datetime.now().isoformat()
-            }
-        except Exception as e:
-            error_str = str(e)
-            # Fallback check for authentication failures from other exception types
-            if 'Unauthorized' in error_str or 'Bad username' in error_str or 'authentication' in error_str.lower():
-                return {
-                    'device': device_name,
-                    'ip': device_ip,
-                    'status': 'unconfigured',
-                    'error': 'Device reachable but authentication failed (not yet configured)',
-                    'last_check': datetime.now().isoformat()
-                }
-            return {
-                'device': device_name,
-                'ip': device_ip,
-                'status': 'error',
-                'error': error_str,
-                'last_check': datetime.now().isoformat()
-            }
-
-    def check_all_devices(self):
-        """Check status of all devices from both modules.yaml and topo_build.yml."""
-        # Get devices from both sources
-        nodes = get_all_devices()
-        statuses = {}
-
-        # Debug logging
-        safe_log('info', f'Found {len(nodes)} devices from all sources', event='api', handler='DeviceStatusAPIHandler')
-
-        # Use thread pool for parallel checks
-        from concurrent.futures import ThreadPoolExecutor, as_completed
-
-        with ThreadPoolExecutor(max_workers=10) as executor:
-            futures = {
-                executor.submit(self.check_device_status, device_name): device_name
-                for device_name in nodes.keys()
-            }
-
-            try:
-                for future in as_completed(futures, timeout=60):
-                    device_name = futures[future]
-                    try:
-                        result = future.result()
-                        statuses[device_name] = result
-                    except Exception as e:
-                        statuses[device_name] = {
-                            'device': device_name,
-                            'status': 'error',
-                            'error': str(e)
-                        }
-            except TimeoutError:
-                # Some futures didn't complete in time — mark remaining as timed out
-                for future, device_name in futures.items():
-                    if device_name not in statuses:
-                        statuses[device_name] = {
-                            'device': device_name,
-                            'status': 'down',
-                            'error': 'Status check timed out',
-                            'last_check': datetime.now().isoformat()
-                        }
-
-        return statuses
-
-
-class RunningConfigAPIHandler(BaseHandler):
-    """API endpoint to fetch running config from a device via eAPI."""
-
-    def get(self):
-        if not self.current_user:
-            self.set_status(401)
-            self.write(json.dumps({'error': 'Authentication required'}))
-            return
-
-        safe_log('info', 'Running config requested', event='api', endpoint='running_config', device=str(self.get_argument('device', 'unknown')))
-        self.set_header("Content-Type", "application/json")
-        self.set_header("Access-Control-Allow-Origin", "*")
-
-        device = self.get_argument('device', None)
-
-        if not device:
-            self.set_status(400)
-            self.write(json.dumps({'error': 'device parameter required'}))
-            return
-
-        try:
-            config = self.get_running_config(device)
-            self.write(json.dumps(config))
-        except Exception as e:
-            safe_log('error', f'Error in RunningConfigAPIHandler: {e}', event='error', handler='RunningConfigAPIHandler')
-            safe_log('error', f'RunningConfigAPIHandler error: {e}', event='api', handler='RunningConfigAPIHandler')
-            traceback.print_exc()
-            self.set_status(500)
-            self.write(json.dumps({'error': str(e)}))
-
-    def get_running_config(self, device_name):
-        """Query EOS device for running config via eAPI."""
-        # Get device IP from topology
-        device_ip = get_device_ip_from_sources(device_name)
-        if not device_ip:
-            raise ValueError(f"Device {device_name} not found in topology")
-
-        # Get credentials from ACCESS_INFO
-        with open(ATD_ACCESS_PATH, 'r') as f:
-            host_yaml = YAML().load(f)
-        username = host_yaml['login_info']['jump_host']['user']
-        password = host_yaml['login_info']['jump_host']['pw']
-
-        # Connect via eAPI
-        try:
-            connection = pyeapi.connect(
-                host=device_ip,
-                username=username,
-                password=password,
-                transport='https',
-                timeout=15
-            )
-
-            # Execute show running-config command with text encoding
-            result = connection.execute(['show running-config'], encoding='text')
-
-            # Get the config output from the text response
-            config_output = result.get('result', [{}])[0].get('output', '')
-
-            return {
-                'device': device_name,
-                'config': config_output,
-                'timestamp': datetime.now().isoformat()
-            }
-
-        except (socket.timeout, OSError) as e:
-            raise ValueError(f"Connection timed out to {device_name} ({device_ip})")
-        except pyeapi.eapilib.ConnectionError as e:
-            # Preserve auth failure info for upstream handler to detect 'unconfigured' status
-            error_str = str(e)
-            if 'Unauthorized' in error_str or 'Bad username' in error_str:
-                raise ValueError(f"Unauthorized: Cannot authenticate to {device_name} ({device_ip}): {e}")
-            raise ValueError(f"Cannot connect to {device_name} ({device_ip}): {e}")
-        except pyeapi.eapilib.CommandError as e:
-            raise ValueError(f"Command error on {device_name}: {e}")
-
-
+# TopologyAPIHandler, DevicesAPIHandler, DeviceTypesAPIHandler,
+# InterfaceStatsAPIHandler, DeviceStatusAPIHandler, RunningConfigAPIHandler
+# — moved to handlers/topology_api.py
 
 # ===============================
 # Packet Capture / Latency / Impairment Handlers
 # Imported from handlers.capture, handlers.latency, handlers.impairments
 # ===============================
 
-class ClientLogHandler(tornado.web.RequestHandler):
-    """Receive client-side log events from browser JS and forward to Cloud Logging."""
-    VALID_LEVELS = {'info', 'warning', 'error'}
-
-    def post(self):
-        try:
-            data = json.loads(self.request.body)
-            level = data.get('level', 'info')
-            if level not in self.VALID_LEVELS:
-                level = 'info'
-            message = str(data.get('message', ''))[:500]
-            source = str(data.get('source', 'unknown'))[:50]
-            action = str(data.get('action', ''))[:50]
-            kwargs = {'event': 'client', 'source': source}
-            if action:
-                kwargs['action'] = action
-            for key in ('device', 'topology', 'session_id', 'client_id'):
-                if key in data:
-                    kwargs[key] = str(data[key])[:100]
-            safe_log(level, message, **kwargs)
-            self.set_status(204)
-        except Exception:
-            self.set_status(204)
-
-
-class ConnectivityStatusHandler(BaseHandler):
-    """REST endpoint returning current connectivity state for live session health"""
-    def get(self):
-        if not self.current_user:
-            self.set_status(401)
-            self.write(json.dumps({'error': 'Authentication required'}))
-            return
-        self.set_header('Content-Type', 'application/json')
-        self.write(json.dumps({
-            'active_sessions': len(active_sessions),
-            'active_session_data': list(active_session_data.values()),
-            'recent_disconnects': len(recent_sessions),
-            'internal_grpc': {
-                'last_status': _last_internal_grpc_status,
-                'last_check': _last_internal_grpc_time
-            }
-        }))
+# ClientLogHandler, ConnectivityStatusHandler — moved to handlers/pages.py
 
 
 if __name__ == "__main__":
@@ -2979,6 +844,34 @@ if __name__ == "__main__":
         'exam_state': EXAM_STATE,
     }
 
+    # Config dict for page handlers — base paths and title
+    _page_config = {
+        'base_path': BASE_PATH,
+        'atd_access_path': ATD_ACCESS_PATH,
+        'title': TITLE,
+    }
+
+    # Topo config dict for topoRequestHandler
+    _topo_config = {
+        'topo': TOPO,
+        'nomenuoptionfile': NOMENUOPTIONFILE,
+        'menu_items': MENU_ITEMS if not NOMENUOPTIONFILE else {},
+        'default_menu_file_value': DEFAULT_MENU_FILE_VALUE if not NOMENUOPTIONFILE else '',
+        'mod_yaml': MOD_YAML,
+        'eos_type': EOS_TYPE,
+    }
+
+    # Session state shared with ConnectivityStatusHandler (mutable references).
+    # active_sessions/active_session_data/recent_sessions are mutated in-place,
+    # so the handler sees live data. _grpc_state is also a mutable dict updated
+    # by check_cvp_grpc_internal.
+    _session_state = {
+        'active_sessions': active_sessions,
+        'active_session_data': active_session_data,
+        'recent_sessions': recent_sessions,
+        'grpc_state': _grpc_state,
+    }
+
     app = tornado.web.Application([
         (r'/td-api/client-log', ClientLogHandler),
         (r'/exam-submitted', ExamSubmittedRedirectHandler, _exam_kwargs),
@@ -2988,7 +881,7 @@ if __name__ == "__main__":
         (r'/css/(.*)', tornado.web.StaticFileHandler, {'path': BASE_PATH +  "css/"}),
         (r'/images/(.*)', tornado.web.StaticFileHandler, {'path': BASE_PATH +  "images/"}),
         (r'/topo/(.*)', tornado.web.StaticFileHandler, {'path': ArBASE_PATH}),
-        (r'/', topoRequestHandler),
+        (r'/', topoRequestHandler, {'config': _page_config, 'topo_config': _topo_config}),
         (r'/td-ws', topoDataHandler),
         (r'/login', LoginHandler, {'accounts': accounts, 'salt': salt, 'base_path': BASE_PATH}),
         (r'/lab', LabHandler, {'docker_client': DOCKER_CLIENT, 'default_menu_file_value': DEFAULT_MENU_FILE_VALUE}),
@@ -3005,10 +898,10 @@ if __name__ == "__main__":
         (r'/getUserSessionId', GetUserSessionIdHandler, _exam_kwargs),
         (r'/beginExam', BeginExamHandler, _exam_kwargs),
         (r'/endExam', EndExamHandler, _exam_kwargs),
-        (r'/baseUrl', BaseUrlHandler),
-        (r'/uptimeWithRuntime', UptimeWithRuntimeHandler),
-        (r'/terminal', TerminalPageHandler),
-        (r'/console/?', ConsolePageHandler),  # /? makes trailing slash optional
+        (r'/baseUrl', BaseUrlHandler, {'config': _page_config}),
+        (r'/uptimeWithRuntime', UptimeWithRuntimeHandler, {'exam_state': EXAM_STATE, 'topo_data': None}),
+        (r'/terminal', TerminalPageHandler, {'config': _page_config}),
+        (r'/console/?', ConsolePageHandler, {'config': _page_config}),  # /? makes trailing slash optional
         (r'/td-api/devices', DevicesAPIHandler),
         (r'/td-api/device-types', DeviceTypesAPIHandler),
         (r'/td-api/topology', TopologyAPIHandler),
@@ -3039,7 +932,7 @@ if __name__ == "__main__":
         (r'/td-api/topology-converter/convert', TopologyConverterConvertHandler),
         (r'/td-api/topology-converter/status', TopologyConverterStatusHandler),
         # Connectivity status endpoint
-        (r'/td-api/connectivity-status', ConnectivityStatusHandler),
+        (r'/td-api/connectivity-status', ConnectivityStatusHandler, {'session_state': _session_state}),
         # Nodebuilder endpoints (dynamic node addition for KVM labs)
         (r'/td-api/nodes/(.*)', NodeBuilderProxyHandler),
     ], **settings)
